Move the new method to a separate service
